Fix string tokenizing+parsing to make a bit more sense
